[STORM-3595] Pull IStrategy classes back into BaseResourceAwareStrategy; introduce ObjectResourceSortType enumeration and further consolidate code.
